Solve compiler warnings

+++ include/base_queue_wait.h
@@ -104,7 +104,7 @@ namespace small {
         inline EnumLock wait_pop(std::vector<T>& vec_elems, int max_count = 1)
         {
             vec_elems.clear();
-            vec_elems.reserve(max_count);
+            vec_elems.reserve(static_cast<std::size_t>(max_count));
 
             std::unique_lock l(m_lock);
